Login: suggest to user to sign in with Google if email/pw is incorrect
This is done because signing in with Google overrides other untrusted auth providers like unverified email, preventing the user from logging in with the original credentials they registered with. The message is displayed if Firebase throws an invalid credentials error which also occurs during an auth provider mismatch (signing in with email/pw when email auth provider isn't available).